refactor: move xor string helper to core security
