+++ internal/il2cpp-dll-injection.vcxproj.filters
@@ -144,6 +144,8 @@
     <ClInclude Include="vendor\imgui\imstb_truetype.h" />
     <ClInclude Include="src\bootstrap\main.h" />
     <ClInclude Include="src\core\config\settings.h" />
+    <ClInclude Include="src\core\security\xorstr.h" />
+    <ClInclude Include="src\ui\xorstr.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="appdata">
